Restructure the benchmarks
So that we can write multiple benchmark groups modules
