+++ Cargo.toml
@@ -39,5 +39,5 @@ criterion = { version = "0.4.0", features = ["html_reports"] }
 rstest = "0.16.0"
 
 [[bench]]
-name = "portgraph"
+name = "bench_main"
 harness = false
