+++ benches/bench_main.rs
@@ -1,0 +1,9 @@
+#[allow(dead_code)]
+mod benchmarks;
+
+use criterion::criterion_main;
+
+criterion_main! {
+    benchmarks::portgraph::portgraph_benches,
+    benchmarks::substitute::substitute_benches,
+}
