+++ benches/benchmarks/mod.rs
@@ -1,0 +1,4 @@
+pub mod generators;
+
+pub mod portgraph;
+pub mod substitute;
